feat: 주석 정리 및 알림 페이지 수정 시작
