+++ src/components/alert/alertList.ts
@@ -1,21 +1,23 @@
 // 알림 더미데이터
 
-export type AlertType =
-  | "invite"
-  | "invite_accept"
-  | "invite_reject"
-  | "change"
-  | "simple";
+// export type AlertType =
+//   | "invite"
+//   | "invite_accept"
+//   | "invite_reject"
+//   | "change"
+//   | "simple";
 
-export interface AlertItem {
-  notificationId: number;
-  title: string;
-  content: string;
-  type: AlertType;
-  isRead: boolean;
-  imageSrc: string;
-  groupId: number; // 모임 이동시 필요
-}
+// export interface AlertItem {
+//   notificationId: number;
+//   title: string;
+//   content: string;
+//   type: AlertType;
+//   isRead: boolean;
+//   imageSrc: string;
+//   groupId: number; // 모임 이동시 필요
+// }
+
+import type { AlertItem } from "../../types/alert";
 
 export const alertList: AlertItem[] = [
   {
@@ -24,7 +26,7 @@ export const alertList: AlertItem[] = [
     content: "07.03 (월) 운동이 수정되었습니다.",
     type: "change",
     isRead: false,
-    imageSrc: "src/assets/images/kitty.png",
+    imgKey: "src/assets/images/kitty.png",
     groupId: 101,
   },
   {
@@ -33,7 +35,7 @@ export const alertList: AlertItem[] = [
     content: "07.03 (월) 운동이 삭제되었어요!",
     type: "simple",
     isRead: false,
-    imageSrc: "src/assets/images/kitty.png",
+    imgKey: "src/assets/images/kitty.png",
     groupId: 101,
   },
   {
@@ -42,7 +44,7 @@ export const alertList: AlertItem[] = [
     content: "'민턴클로버' 모임에 초대받았습니다.",
     type: "invite",
     isRead: false,
-    imageSrc: "src/assets/images/kitty.png",
+    imgKey: "src/assets/images/kitty.png",
     groupId: 101,
   },
 ];
